+++ cdoc2-cli/src/main/java/ee/cyber/cdoc2/cli/commands/CDocListCmd.java
@@ -70,7 +70,7 @@ public class CDocListCmd implements Callable<Void> {
         }
 
         DecryptionKeyMaterial decryptionKeyMaterial = (null == this.exclusive)
-            ? getSmartCardDecryptionKeyMaterial(this.slot, this.keyAlias)
+            ? getSmartCardDecryptionKeyMaterial(this.slot, this.keyAlias, null)
             : getDecryptionKeyMaterial(this.cdocFile, this.exclusive);
 
         CDocDecrypter cDocDecrypter = new CDocDecrypter()