+++ cdoc2-lib/src/main/java/ee/cyber/cdoc2/CryptoStickConf.java
@@ -1,0 +1,19 @@
+package ee.cyber.cdoc2;
+
+/**
+ * Enum holding information about the crypto stick used for decryption
+ */
+public enum CryptoStickConf {
+    SECP256R1(256),
+    SECP384R1(384);
+
+    private final int keySize;
+
+    CryptoStickConf(int keySize) {
+        this.keySize = keySize;
+    }
+
+    public int getKeySizeInBytes() {
+        return keySize / 8;
+    }
+}